Safari-611.1.10/LayoutTests/accessibility/mac/content-editable-range-properties-expected.txt

Single line test
Multiple
Line
Textbox
This tests that contenteditable nodes return the correct information for range and selection based attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axElement.selectedTextRange is '{3, 0}'
PASS axElement.selectedTextRange is '{3, 3}'
PASS axElement.stringAttributeValue('AXSelectedText') is 'gle'
PASS axElement.selectedTextRange is '{3, 4}'
PASS axElement.selectedTextRange is '{12, 0}'
PASS axElement.selectedTextRange is '{12, 6}'
PASS axElement.stringAttributeValue('AXSelectedText') is matchedText
PASS axElement.selectedTextRange is '{12, 7}'
PASS successfullyParsed is true

TEST COMPLETE
